fix(detail): only show "more" when description actually overflows
ExpandableText offered the more/less toggle whenever the text exceeded
160 characters, a heuristic that is wrong on wide layouts where a longer
synopsis still fits within the collapsed line limit. The button then did
nothing visible when tapped.

Measure the rendered text instead — compare a collapsed-clamped copy
against an unbounded one (the same probe TVAboutText already uses on
tvOS) and only show the toggle when the full height genuinely exceeds
the collapsed height. Split ExpandableText into its own file to keep
MediaDetailComponents under the file-length limit.

Fixes #83
